requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
azure-cognitiveservices-speech>=1.40.0
httpx>=0.28.0
